Add ConsoleIo to the container for commands
I have a todo from a discussion somewhat recently about adding ConsoleIo
and Arguments to the container so that they could be injected into the
Command as attributes.

I've done that for ConsoleIo. However, doing it for `Arguments` is not
possible. The reason it isn't possible is that `Arguments` requires the
`ConsoleOptionParser` instance from the command, which creates a cyclic
dependency. We _could_ pass an empty `Arguments` in but I don't see how
that is anything but a footgun.